I'm not sure what to do.
I'm wondering. Why is it? Why is this happening in the new version of the compiler?

Were we very close to the edge in the previous version. And the changes in the latest one only pushed it over. Or is it something else?

I want to go to the earlier version of the compiler and seeing whether I can get it to crash.