I keep wanting for somebody or something to save me from the fate of having to go forward with this, but that is not to be.
We're going to have to run this.

So the next step would be to integrate this with the UI.

We have no choice to go forward even if what awaits us is destruction.

We implemented everything that we need on the backend.

After the break we'll take a look at the UI.