"Even though I made sure to separate out the shared memory bein used using barriers, it is allocating every shared array separately.
This is a huge issue.

The design of the whole library is broken right now.

I am going to first open an issue on the Nvidia dev support page to complain, and in the next video, we're going to redesign all of this so it only uses dynamic shared memory."