"The ensembles are at around 1.8 positive.
This is how I remember the successful run of tabular CFR being on the host against a random player.

If we turn off the updates..

Weird that it's slightly positive.

But I guess that reflects all the times pl1 raises and the other folds.

Rando vs rando, pl1 has a slight advantage vs pl2.

I am astonished.

Amazed.

We implemented all this...and the training actually worked the first time?

Insane."