"Let's stop here for the day.
We'll check out the reg count for this tomorrow.

Honestly, if this isn't a way to optimize the reg counts, we're pretty screwed.

Even so, I believe my logic is sound, so we should get through this.

If we can...

I guess the next thing would be to check out Cutlass and work more on matrix multiplication."