"Sigh, now it starts at 136.
No wait, it'd be more fair to say it starts at 7.

Here it is storing a bunch of data into local memory.

What if we backed up the table?

I have no idea what is going on here anymore.

Was there a point to any of these optimizations?

...

I am going to pause here and revisit the old commits.

I'll go back before we did the layout type modifications.

..."