https://github.com/NVIDIA/cutlass/issues/1811
I never expected to run into a bug like this.

I guess I'll have to stop trying to get it to run with CuPy, and switch to trying to setup a hello world example that uses the Cutlass kernel API from a C++ project.

I wanted to avoid having to do direct C++ programming since Spiral would have made it easy to make test examples thanks to its tensors, but Cutlass is proving as hard to get off the ground as I had expected.