Every time I try out Cutlass it feels like I waste nearly a week before giving up. I am depressed that I'll be doing a matrix multiply on my own again. Anyway, I'll be investigating async loads again.
Here we have an example that mostly writes and reads from memory.

If I add that `nanosleep 300` call, in Nsight Compute the memory occupancy drop from 92% to 37% and the runtime goes from 8ms to 23ms.

Ordinarily there'd be no way to speed this up, but if I interleaved memory loads with computation, I should be able to double the memory occupancy...