Finally got something usable. The sync version works best, probably due to being able to use large tiles.
I think the crazy version only worked so due to the some lucky compiler optimizations. As far as I can tell it's not good to convert to the tf32 format in transit. It's better to convert after loading from shared into local even if it takes more computation than having an intermediate step.