Fixed the player selection bug.
Also, done with the transpose. Needs to be tested.

But I don't like this version so we'll change it so it uses a simple projective loop instead. The problem with the current approach is that it won't support arbitrary dimensions, and I do not want to restrict this kernel to supporting only dims divisible by the number of threads per block.