Making further progress. Now the softmax is causing trouble. The CEM method might be causing the weights to grow unbounded.
"Got it. Yeah, this is the issue.

Once I put this here the assert gets triggered.

As to why this is happening once we started doing reward resetting, I can only guess that it is due to more weight drifting happening.

Before, not doing the weight resetting might have allowed the old winners to stick around for longer.

But I am going to investigate this more deeply to make sure that this is in fact what is going on.

We'll also have to figure out what to do about the softmax.

It doesn't seem like it works too well with the CEM.

We'll take a look at the weight norms and decide after seeing whether they are blowing up.

We'll stop here for the day."