Just looking into how strings can be mapped in PS. The way I am going to manage deps is have the script download the needed Github repos for libs like Thunderkitters and Crow if the dirs for them don't exist and then pass those dirs as includes to nvcc.
I know the principle is to use a package manager and an existing build system rather than making an ad hoc script for the same thing, but at a certain level of complexity it starts to make sense.

With my own build script I can follow the steps easily and do anything I want, but with Conan + CMake I'd be messing around with build systems for a long time. I have better things to do. I hate dealing with config files most of all!