5:20pm. https://docs.nvidia.com/cuda/cuda-gdb/index.html#supported-platforms
Trying to debug Cuda device code in WSL and it turns out I did have to set the `HKEY_LOCAL_MACHINE\SOFTWARE\NVIDIA Corporation\GPUDebugger\EnableInterface` registry key to `(DWORD 32-bit) 1`.

Unfortunately, my machine nearly crashed just now.

5:40pm. I realized why it's being unstable.

```cpp
int main_body() {
    thrust::device_vector<int> v0 = {1,2,3,4};
    int v1;
    v1 = v0.size();
    int v2;
    v2 = v0[2];
    printf("{%s = %d; %s = %d; %s = %s}\n","index_2", v2, "length_of_array", v1, "message", "hello from host");
    run_cuda_host_0(v0);
    return 0;
}
int main(){
    auto r = main_body();
    gpuErrchk(cudaDeviceSynchronize()); // This line is here so the `__trap()` calls on the kernel aren't missed.
    return r;
}
```

It's because the `v0` device vector is getting freed before the kernel finishes running. When running the program normally it doesn't seem to be a problem, but when combined with a debugger it leads to system instability and screen flickering.

```cpp
int main_body() {
    thrust::device_vector<int> v0 = {1,2,3,4};
    int v1;
    v1 = v0.size();
    int v2;
    v2 = v0[2];
    printf("{%s = %d; %s = %d; %s = %s}\n","index_2", v2, "length_of_array", v1, "message", "hello from host");
    run_cuda_host_0(v0);
    gpuErrchk(cudaDeviceSynchronize()); // This line is here so the `__trap()` calls on the kernel aren't missed.
    return 0;
}
int main(){
    auto r = main_body();
    return r;
}
```

If I move the device synchronize a few lines up like this so that the `v0`'s destructor gets called after the kernel is done, the problems go away.

5:50pm. I guess I'll have to remove it from the code generator. I'll also remove them from the Python backend as well.

Indeed, I'd very prefer for this to have been inserted into the main function, but I couldn't think of a good way to do it.

I guess having to put them in manually wouldn't be too much of a deal.