6:25pm.The latest changes to the compiler work. Will continue next week. My plan then is to try going through some of the Cutlass examples with the debugger.
My battle with that library is far from over. If I could step through it with a debugger, I could properly study it which would give me a lot of insight on GPU performance and lead to the creation of Hopfield dictionaries.

Right now I am busy with my paid work, so I don't have time for this project, but in at most few years I'll hit retirement and get back to this. It really depends how the company I am with does.

It could be in as little as 6 months, it could be as long as half a decade.

It also depends on how my trading account fares.

These initial months matter the most to me in the financial sense, and hopefully I can make the project we're working on a success. I am pretty happy with the job so far despite some hickups along the way.